pyproject.toml
[tool.poetry]
name = "repo-size"
version = "0.1.0"
description = ""
authors = ["Mischa <mishayartsev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.25.2"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
